Fixed errors when cp nginx.conf

+++ install_nextcloud.sh
@@ -147,7 +147,7 @@ sleep 1
 ##################################
 #Create nextcloud.conf file in /etc/nginx/conf.d/ directory"
 ##################################
-mv nginx.conf nginx_conf_backup
+mv /etc/nginx/nginx.conf /etc/nginx/nginx_conf_backup
 cp $HOME/ubuntu-lnmp-nextcloud-setup/nginx.conf /etc/nginx/nginx.conf
 if [ -f /etc/nginx/conf.d/default.conf ]; then
         mv /etc/nginx/conf.d/default.conf /etc/nginx/conf.d/default_conf_disabled
@@ -165,6 +165,8 @@ if [ ! -f /etc/nginx/sites-available/nextcloud.conf ]; then
         cp $HOME/ubuntu-lnmp-nextcloud-setup/nextcloud_HTTP_nginx.conf /etc/nginx/sites-available/nextcloud.conf
         #Now create a symbolic link from nextcloud block configuration file to the /etc/nginx/sites-enabled/ directory:
         ln -s /etc/nginx/sites-available/nextcloud.conf /etc/nginx/sites-enabled/
+fi
+if [-f /etc/nginx/sites-enabled/default ]; then
         #Unlink the default configuration
         sudo unlink /etc/nginx/sites-enabled/default
 fi
